Complete rewrite
Inversion of control. Instead of passing in a big configuration object
to some generic code, each site is now implemented with a custom bit of
code. However, the helper functions in the totally new Blinder object
make this code very concise, easy to read *and* to write.

This should also make it easier to adapt to other kind of page
structures without having to make some generic top-level code
increasingly complicated.

Needs documentation.